[style] #153 리뷰 반영 - Long 확장함수 추가
