+++ presentation/src/main/java/com/mashup/presentation/feature/home/model/SignalUiModel.kt
@@ -3,7 +3,7 @@ package com.mashup.presentation.feature.home.model
 import com.mashup.domain.model.ReceivedSignal
 import com.mashup.presentation.common.base.UiMapper
 import com.mashup.presentation.common.base.UiModel
-import java.util.concurrent.TimeUnit
+import com.mashup.presentation.common.extension.getDisplayedTime
 
 data class SignalUiModel(
     val signalId: Int = -1,
@@ -16,33 +16,6 @@ data class SignalUiModel(
     val keywordsCount: Int = -1,
     val receivedTimeMillis: String = "",
 ) : UiModel, UiMapper<ReceivedSignal, SignalUiModel> {
-    fun getDisplayedTime(receivedTimeMillis: Long): String {
-        val currentTimeMillis = System.currentTimeMillis()
-        val diffMillis = currentTimeMillis - receivedTimeMillis
-
-        return when {
-            diffMillis < TimeUnit.MINUTES.toMillis(1) -> "지금"
-            diffMillis < TimeUnit.HOURS.toMillis(1) -> {
-                val minutes = TimeUnit.MILLISECONDS.toMinutes(diffMillis)
-                "${minutes}분 전"
-            }
-            diffMillis < TimeUnit.DAYS.toMillis(1) -> {
-                val hours = TimeUnit.MILLISECONDS.toHours(diffMillis)
-                "${hours}시간 전"
-            }
-            diffMillis < TimeUnit.DAYS.toMillis(30) -> {
-                val days = TimeUnit.MILLISECONDS.toDays(diffMillis)
-                "${days}일 전"
-            }
-            diffMillis < TimeUnit.DAYS.toMillis(365) -> {
-                val months = (TimeUnit.MILLISECONDS.toDays(diffMillis) / 30).toInt()
-                "${months}달 전"
-            }
-            else -> {
-                "오래 전"
-            }
-        }
-    }
     override fun toUiModel(domain: ReceivedSignal) = SignalUiModel(
         signalId = domain.signalId,
         receiverId = domain.receiverId,
@@ -52,6 +25,6 @@ data class SignalUiModel(
         signalContent = domain.signalContent,
         keywords = domain.keywords.map { "#$it" },
         keywordsCount = domain.keywordsCount,
-        receivedTimeMillis = getDisplayedTime(domain.receivedTimeMillis)
+        receivedTimeMillis = domain.receivedTimeMillis.getDisplayedTime()
     )
 }
